Move extract-insee-code to helpers
